Consolidate Note and Memo into NoteAndMemo
